fixed the error

+++ 0x0A-configuration_management/2-execute_a_command.pp
@@ -2,6 +2,6 @@
 
 exec {'killmenow':
 command => 'pkill killmenow',
-only    => 'pgrep killmenow',
+onlyif  => 'pgrep killmenow',
 path    => ['/bin/'],
 }
